Add generic imaging option in new request and new imaging request forms
Add in a new table column in each image resolution sub-form for generic
imaging so that the user can get their images without committing to
additional analysis. Add a column in the db table to hold the boolean
for generic_imaging, like all of the other imaging modes. Also fix
an issue where the button for making a new imaging resolution table
in the new imaging request form was raising a validation error
that should only have occurred if the button pressed was the final
submit button.