Introduce new database log table for user actions
In recent commits, I attempted to use Datajoint's Log()
table to log my user events. However, this table uses
a timestamp as the primary key and this timestamp only
goes down to the second. This is undesirable for
logging user requests which can happen many times
per second. To solve this, I created a new table
in the light sheet schema called UserActionLog()
which will log the user events using an integer
called "event_number" as the primary key.
This is an autoincrement integer so it will
simply add an each log entry even if it is
entered in during the same second as a previous
entry.