add schedulers to poll statuses of raw and tiled precomputed pipelines
Add celery tasks which will run as schedules to poll when all
channels are done being converted to precomputed format.
When this happens send an email to the user with a link
to a visualization setup form. This is done for both
raw data (single tile) precomputed pipelines as well as
tiled data precomputed pipelines. Routes and forms
are written for setting up the neuroglancer viewer
based on which channels/lightsheets the user wants to display.

The actual launching of the cloudvolumes is pending
because their containers must be launched as root,
but the lightservuser explicitly cannot be root
in order to write to bucket.