add new flask app 'viewer-launcher' to handle launching docker containers from lightserv
This commit gets around the issue from the previous commit where
containers for cloudvolumes and neuroglancer viewers could
not be launched from lightserv because I was not root in that container.
The 'viewer-launcher' app runs on the same network as lightserv
but runs as root so it can launch containers. 'viewer-launcher'
is a simple API that lightserv sends POSTs to to launch the containers.

This commit also adds a really simple template for showing the
neuroglancer viewer link.