complete registration precomputed pipelines
add ability to launch cloudvolumes for multiple channels
that were registered to an atlas during the light sheet
processing pipeline. If an atlas was selected in the
visualization setup form, then an atlas cloudvolume is created,
but only once to save memory. The atlas is overlaid
on any channel for which it was requested in the form